+++ src/ui/first_run/dependencies.rs
@@ -66,7 +66,7 @@ impl SimpleAsyncComponent for DependenciesApp {
                         },
 
                         gtk::Entry {
-                            set_text: "sudo pacman -Syu git p7zip libwebp-utils",
+                            set_text: "sudo pacman -Syu git p7zip",
                             set_editable: false
                         }
                     },
@@ -83,7 +83,7 @@ impl SimpleAsyncComponent for DependenciesApp {
                         },
 
                         gtk::Entry {
-                            set_text: "sudo apt install git p7zip-full webp",
+                            set_text: "sudo apt install git p7zip-full",
                             set_editable: false
                         }
                     },
@@ -100,7 +100,7 @@ impl SimpleAsyncComponent for DependenciesApp {
                         },
 
                         gtk::Entry {
-                            set_text: "sudo dnf install git p7zip libwebp-tools",
+                            set_text: "sudo dnf install git p7zip",
                             set_editable: false
                         }
                     },
@@ -119,10 +119,6 @@ impl SimpleAsyncComponent for DependenciesApp {
 
                             adw::ActionRow {
                                 set_title: "p7zip"
-                            },
-
-                            adw::ActionRow {
-                                set_title: "libwebp"
                             }
                         }
                     }
@@ -196,7 +192,7 @@ impl SimpleAsyncComponent for DependenciesApp {
         match msg {
             #[allow(unused_must_use)]
             DependenciesAppMsg::Continue => {
-                let packages = ["git", "dwebp"];
+                let packages = ["git"];
 
                 for package in packages {
                     if !is_available(package) {