+++ CHANGELOG.md
@@ -21,6 +21,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Changed mfc140 installation method to not use winetricks
 
+### Removed
+
+- Removed dependency on dwebp/libwebp-utils
+
 ## [1.14.4] - 2026-02-13
 
 ### Added